integrations/llamaindex/pyproject.toml
[project]
name = "neuralmem-llamaindex"
version = "0.2.0"
description = "LlamaIndex integration for NeuralMem — BaseRetriever and BaseMemory adapters"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "NeuralMem Team" }]
dependencies = [
    "neuralmem>=0.2.0",
    "llama-index-core>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["neuralmem_llamaindex"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short -q"
